requirements.txt
ebaysdk == 2.1.5
PyYAML == 3.13
Jinja2 == 2.10
sendgrid == 5.6.0
click == 7.0
